+++ src/Burkus.Mvvm.Maui/Attributes/MapNavigationParameterAttribute.cs
@@ -11,11 +11,21 @@ public class MapNavigationParameterAttribute : Attribute
 
     public string NavigationParameterKey { get; }
 
+    public bool Required { get; }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="propertyName">The property name on the viewmodel to map to</param>
+    /// <param name="navigationParameterKey">The navigation parameter to map from</param>
+    /// <param name="required">Throws an exception if required is set to false</param>
     public MapNavigationParameterAttribute(
         string propertyName,
-        string navigationParameterKey)
+        string navigationParameterKey,
+        bool required = false)
     {
         PropertyName = propertyName;
         NavigationParameterKey = navigationParameterKey;
+        Required = required;
     }
 }